+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased
 Deprecated
  - `tcod.console_load_xp` has been deprecated, `tcod.console_from_xp` can load
    these files without modifying an existing console.
+ - Support for Python 3.5 will be dropped soon.
 
 Fixed
  - `tcod.console_from_xp` now has better error handling (instead of crashing.)